app_backend/requirements.txt
Flask==2.0.4
Flask-SQLAlchemy==3.0.1
flask-cors==3.1.1
flask-uploads==0.2.1
python-dotenv==0.19.5
redis==4.1.1
requests==2.26.0
schedule==1.1.0
smtplib==3.9.0
Flask-RESTful==0.3.9
PyJWT==2.3.0
